restructured and updated file paths accordingly.
