+++ tests/test_call_command.py
@@ -1,3 +1,3 @@
 import subprocess
 
-subprocess.run(["python", "../sagittal_brain.py"], shell=True)
+subprocess.run(["python", "../src/sagittal_brain/sagittal_brain.py"], shell=True)
